Rewrite x86 Unwinder using UnwindStackFrame
