+++ src/unwinder/CMakeLists.txt
@@ -17,12 +17,6 @@ list(APPEND UNWINDER_SOURCES
     ${ARCH_SOURCES_DIR}/unwinder_${ARCH_SOURCES_DIR}.cpp
 )
 
-if(CLR_CMAKE_PLATFORM_ARCH_I386 OR CLR_CMAKE_PLATFORM_ARCH_AMD64)
-list(APPEND UNWINDER_SOURCES
-    unwinder_xarch.cpp
-)
-endif(CLR_CMAKE_PLATFORM_ARCH_I386 OR CLR_CMAKE_PLATFORM_ARCH_AMD64)
-
 convert_to_absolute_path(UNWINDER_SOURCES ${UNWINDER_SOURCES})
 
 if(CLR_CMAKE_PLATFORM_UNIX)
